src/requirements.txt
boto3
python-dotenv
pyperclip
json
sys
os
spacy
